Moved implementations; use j.t.Duration in Java API
